fix(greatclips): cross-platform CookieHeader + empty patch index for CI
