+++ library/other/greatclips/.printing-press-patches.json
@@ -1,0 +1,4 @@
+{
+  "schema_version": 1,
+  "patches": []
+}
